# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Example api Documentation"
        }
    ],
    "paths": {
        "/tweets": {
            "get": {
                "tags": [
                    "Tweet"
                ],
                "summary": "Tweets List.",
                "operationId": "tweet-list",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tweet"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid tag value"
                    }
                }
            }
        },
        "/tweets/{id}": {
            "get": {
                "tags": [
                    "Tweet"
                ],
                "summary": "Tweet By ID.",
                "operationId": "tweet-by-id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "tweet id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tweet"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tweet"
                ],
                "operationId": "tweet-by-id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "tweet id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pet object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Tweet"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tweet"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tweet"
                ],
                "summary": "Remove Tweet By ID.",
                "operationId": "tweet-by-id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "tweet id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Tweet": {
                "required": [
                    "fullname",
                    "username",
                    "avatar",
                    "quote"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "fullname": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "avatar": {
                        "type": "string"
                    },
                    "quota": {
                        "type": "string"
                    },
                    "likes": {
                        "type": "integer"
                    },
                    "retweets": {
                        "type": "integer"
                    },
                    "comments": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "date-time"
                    },
                    "updated_at": {
                        "type": "date-time"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Tweet"
                }
            }
        }
    }
}